+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## v2.1.0
--   Add new component for `<pxb-user-menu` (use @pxblue/angular-themes v5.1.0+ to get PX Blue themes for this component).
+-   Add new component for `<pxb-user-menu>` (use @pxblue/angular-themes v5.1.0+ to get PX Blue themes for this component).
 
 
 ## v2.0.0